+++ display.c
@@ -1157,7 +1157,7 @@ void whatCursor(void) {
 	}
 
 	/* Get character at cursor */
-	/* TODO Unicode codepoint uint32_t cp = utf8ToUCS(row->chars, E.buf->cx); */
+	/* TODO Unicode codepoint uint32_t cp = utf8Decode(row->chars, E.buf->cx); */
 	char ch[8] = "EOL";
 	if (E.buf->cy < E.buf->numrows &&
 	    E.buf->cx < E.buf->row[E.buf->cy].size) {